requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.5
django-bootstrap4==3.0.1
django-smart-selects==1.5.9
djangorestframework==3.12.4
Pillow==8.3.1
psycopg2-binary==2.9.1
python-dotenv==0.18.0
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
